refactor: template processing
