+++ lib/index.ts
@@ -2,13 +2,37 @@ import * as fs from "fs";
 import * as path from "path";
 import * as handlebars from "handlebars";
 
-export const processTemplates = async (
-  templateDir: string,
+const prepareData = (args: any) => {
+  const argsList = args.arguments || [];
+  const names = argsList.map((i: string) => i.split(":")[0]);
+  const types = argsList.map((i: string) => {
+    let parts = i.split(":");
+    // If there's a type and it's not empty, use it; if not, default to "bytes32"
+    let t =
+      parts.length > 1 && parts[1].trim() !== "" ? parts[1].trim() : "bytes32";
+    return t;
+  });
+  const pairs = names.map((v: string, i: string) => [v, types[i]]);
+  const contractName = sanitizeSolidityFunctionName(args.name);
+
+  return {
+    args,
+    contractName,
+    contractNameUnderscore: camelToUnderscoreUpper(contractName),
+    arguments: { names, types, pairs },
+  };
+};
+
+const processTemplatesRecursive = async (
+  template: string,
   outputDir: string,
   data: Record<string, unknown>
 ): Promise<void> => {
   try {
-    templateDir = path.resolve(__dirname, templateDir);
+    const templateDir = path.resolve(
+      __dirname,
+      path.resolve(__dirname, "..", "templates", template)
+    );
 
     const files = fs.readdirSync(templateDir);
 
@@ -26,7 +50,7 @@ export const processTemplates = async (
 
       if (fs.lstatSync(templatePath).isDirectory()) {
         // If file is a directory, recursively process it
-        await processTemplates(templatePath, outputPath, data);
+        await processTemplatesRecursive(templatePath, outputPath, data);
       } else if (path.extname(file) === ".hbs") {
         const templateContent = fs.readFileSync(templatePath, "utf-8");
         const template = handlebars.compile(templateContent);
@@ -41,11 +65,19 @@ export const processTemplates = async (
   }
 };
 
-export const camelToUnderscoreUpper = (input: string): string => {
+export const processTemplates = async (templateName: string, args: any) => {
+  processTemplatesRecursive(
+    templateName,
+    path.resolve(process.cwd()),
+    prepareData(args)
+  );
+};
+
+const camelToUnderscoreUpper = (input: string): string => {
   return input.replace(/([a-z])([A-Z])/g, "$1_$2").toUpperCase();
 };
 
-export const sanitizeSolidityFunctionName = (str: string): string => {
+const sanitizeSolidityFunctionName = (str: string): string => {
   // Remove any character that's not alphanumeric or underscore
   const cleaned = str.replace(/[^a-zA-Z0-9_]/g, "");
