+++ tasks/omni.ts
@@ -2,29 +2,10 @@ import { task } from "hardhat/config";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 import * as fs from "fs";
 import * as path from "path";
-import { processTemplates, sanitizeSolidityFunctionName } from "../lib";
+import { processTemplates } from "../lib";
 
 const main = async (args: any, hre: HardhatRuntimeEnvironment) => {
-  const templateDir = path.resolve(__dirname, "..", "templates", "omnichain");
-  const outputDir = path.resolve(process.cwd());
-  const argsList = args.arguments || [];
-  const names = argsList.map((i: string) => i.split(":")[0]);
-  const types = argsList.map((i: string) => {
-    let parts = i.split(":");
-    // If there's a type and it's not empty, use it; if not, default to "bytes32"
-    let t =
-      parts.length > 1 && parts[1].trim() !== "" ? parts[1].trim() : "bytes32";
-    return t;
-  });
-  const pairs = names.map((v: string, i: string) => [v, types[i]]);
-
-  const data = {
-    args,
-    contractName: sanitizeSolidityFunctionName(args.name),
-    arguments: { names, types, pairs },
-  };
-
-  processTemplates(templateDir, outputDir, data);
+  processTemplates("omnichain", args);
 
   const configPath = path.resolve(process.cwd(), "hardhat.config.ts");
   let hardhatConfigContents = fs.readFileSync(configPath, "utf8");
