+++ templates/messaging/tasks/interact.ts.hbs
@@ -25,7 +25,7 @@ Please, refer to ZetaChain's explorer for updates on the progress of the cross-c
 `);
 };
 
-task("message", "Sends a message from one chain to another.", main)
+task("interact", "Sends a message from one chain to another.", main)
   .addParam("contract", "Contract address")
   .addParam("amount", "Token amount to send")
   .addParam("destination", "Destination chain ID (integer)");
